+++ expyriment/io/_screen.py
@@ -38,12 +38,12 @@ class Screen(Output):
     """
 
     def __init__(self, colour, open_gl, window_mode, window_size, no_frame,
-                 display):
+                 display, display_resolution):
         """Create and set up a screen output.
 
         Notes
         -----
-        CAUTION: We discourage from creating a screen instance manually!
+        CAUTION: We discourage creating a screen instance manually!
 
         Parameters
         ----------
@@ -53,18 +53,18 @@ class Screen(Output):
             0/False - No OpenGL (no vsync / no blocking)
             1       - OpenGL (vsync / no blocking)
             2/True  - OpenGL (vsync / blocking)
-            3       - OpenGL (vsync / alternative blocking)
         window_mode : bool
-             if screen should be a window
+             if screen should be a window, fullscreen otherwise
         window_size : (int, int)
             size of the window in window_mode,
-            full screen mode if size of window_mode[0]<=0
         no_frame : bool
             set True for windows (in window mode) with no frame;
             this parameter does not affect fullscreen mode
         display : int
             the display index to show the screen on (only has an effect if
             Pygame version >= 2)
+        display_resolution : list or None
+            the resolution of the display
 
         """
 
@@ -73,6 +73,12 @@ class Screen(Output):
         if open_gl == False:
             open_gl = 0
         elif open_gl == True:
+            open_gl = 2
+        elif open_gl >= 2:
+            warn_message = "OpenGL mode '{0}' does not exist. \
+                OpenGL will be set to '2' (default)".format(open_gl)
+            print("Warning: " + warn_message)
+            _internals.active_exp._event_file_warn("Screen,warning," + warn_message)
             open_gl = 2
         self._open_gl = open_gl
         self._fullscreen = not window_mode
@@ -89,66 +95,65 @@ OpenGL will be deactivated!"
 
         pygame.display.init()
         if _internals.active_exp.is_initialized:
-            self._monitor_resolution = \
-                        _internals.active_exp.screen.monitor_resolution
+            self._display_resolution = \
+                        _internals.active_exp.screen.display_resolution
         else:
-            self._monitor_resolution = (pygame.display.Info().current_w,
-                                        pygame.display.Info().current_h)
-        if self._fullscreen:
-            self._window_size = self._monitor_resolution
-        else:
-            self._window_size = window_size
+            if display_resolution is not None:
+                self._display_resolution = display_resolution
+            else:
+                self._display_resolution = pygame.display.list_modes(
+                    display=display)[0]
 
         # Change icon to Expyriment logo
         if platform.system() == "Windows":
             id_ = 'mycompany.myproduct.subproduct.version'  # arbitrary string
             ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(id_)
-        icon = pygame.image.load(
-            os.path.join(os.path.split(__file__)[0], "..", "xpy_icon.png"))
+        icon = pygame.image.load(os.path.join(os.path.split(__file__)[0],
+                                              "..", "xpy_icon.png"))
+        icon = pygame.transform.smoothscale(icon, (32, 32))
         pygame.display.set_icon(icon)
 
         if not self._open_gl:
             if self._fullscreen:
-                if int(pygame.version.ver[0]) > 1:
-                    self._surface = pygame.display.set_mode(
-                        self._window_size, pygame.FULLSCREEN,
-                        display=self._display)
-                else:
-                    self._surface = pygame.display.set_mode(self._window_size,
-                                                            pygame.FULLSCREEN)
+                self._surface = pygame.display.set_mode(
+                    self._display_resolution, pygame.FULLSCREEN,
+                    display=self._display)
             else:
-                if int(pygame.version.ver[0]) > 1:
-                    self._surface = pygame.display.set_mode(
-                        self._window_size, display=self._display)
-                else:
-                    self._surface = pygame.display.set_mode(self._window_size)
+                self._surface = pygame.display.set_mode(
+                    self._window_size, display=self._display)
                 pygame.display.set_caption('Expyriment')
 
         else:
+            os.environ["SDL_RENDER_DRIVER"] = "opengl"
+            os.environ["SDL_FRAMEBUFFER_ACCELERATION"] = "opengl"
+            os.environ["SDL_RENDER_VSYNC"] = "1"
+            os.environ["SDL_HINT_RENDER_BATCHING"] = "0"
+
             try:
                 pygame.display.gl_set_attribute(pygame.GL_SWAP_CONTROL, 1)
             except Exception:
                 pass
 
             pygame_mode = pygame.DOUBLEBUF | pygame.OPENGL
+
+            if platform.system() == "Windows":
+                user32 = ctypes.windll.user32
+                user32.SetProcessDPIAware()
+
             if self._fullscreen:
-                if int(pygame.version.ver[0]) > 1:
-                    self._surface = pygame.display.set_mode(
-                        self._window_size, pygame_mode | pygame.FULLSCREEN,
-                        display=self._display, vsync=1)
-                else:
-                    self._surface = pygame.display.set_mode(
-                        self._window_size, pygame_mode | pygame.FULLSCREEN)
+                os.environ["SDL_VIDEO_X11_NET_WM_BYPASS_COMPOSITOR"] = "1"
+                os.environ["SDL_VIDEO_MAC_FULLSCREEN_SPACES"] = "0"
+
+                self._surface = pygame.display.set_mode(
+                    self._display_resolution, pygame_mode | pygame.FULLSCREEN,
+                    display=self._display, vsync=1)
+
             else:
                 if no_frame:
                     pygame_mode = pygame_mode | pygame.NOFRAME
-                if int(pygame.version.ver[0]) > 1:
-                    self._surface = pygame.display.set_mode(
-                        self._window_size, pygame_mode, display=self._display,
-                        vsync=1)
-                else:
-                    self._surface = pygame.display.set_mode(
-                        self._window_size, pygame_mode)
+                self._surface = pygame.display.set_mode(
+                    self._window_size, pygame_mode, display=self._display,
+                    vsync=1)
 
                 pygame.display.set_caption('Expyriment')
 
@@ -210,10 +215,20 @@ machine!")
         return self._window_size
 
     @property
-    def monitor_resolution(self):
-        """Getter for monitor_resolution."""
+    def display_resolution(self):
+        """Getter for display_resolution."""
 
-        return self._monitor_resolution
+        return self._display_resolution
+
+    @property
+    def monitor_resolution(self):
+        """Getter for monitor_resolution.
+
+        DEPRECATED! Use display_resolution instead.
+
+        """
+
+        return self._display_resolution
 
     def update(self):
         """Update the screen.
@@ -225,11 +240,10 @@ machine!")
         pygame.event.pump()
         pygame.display.flip()
         if self._open_gl >= 2:
-            if self._open_gl == 3:
-                ogl.glBegin(ogl.GL_POINTS)
-                ogl.glColor4f(0, 0, 0, 1)  # Opaque needed for non-alpha video!
-                ogl.glVertex2i(0, 0)
-                ogl.glEnd()
+            ogl.glBegin(ogl.GL_POINTS)
+            ogl.glColor4f(0, 0, 0, 1)  # Opaque needed for non-alpha video!
+            ogl.glVertex2i(0, 0)
+            ogl.glEnd()
             ogl.glFinish()
         if self._logging:
             _internals.active_exp._event_file_log("Screen,updated", 2)
@@ -274,12 +288,12 @@ machine!")
 
         Notes
         -----
-        Each initialized experiment has its one screen (exp.screen). Please use always the
-        screen of your current  experiment.
+        Each initialized experiment has its one screen (exp.screen).
+        Please use always the screen of your current experiment.
 
         """
 
-        return self._window_size[0] // 2
+        return self.size[0] // 2
 
     @property
     def center_y(self):
@@ -287,12 +301,12 @@ machine!")
 
         Notes
         -----
-        Each initialized experiment has its one screen (exp.screen). Please use always the
-        screen of your current  experiment.
+        Each initialized experiment has its one screen (exp.screen).
+        Please use always the screen of your current experiment.
 
         """
 
-        return self._window_size[1] // 2
+        return self.size[1] // 2
 
     @property
     def size(self):
@@ -300,12 +314,12 @@ machine!")
 
         Notes
         -----
-        Each initialized experiment has its one screen (exp.screen). Please use always the
-        screen of your current  experiment.
+        Each initialized experiment has its one screen (exp.screen).
+        Please use always the screen of your current experiment.
 
         """
 
-        return self._window_size
+        return self._surface.get_size()
 
     def clear(self):
         """Clear the screen.
@@ -335,4 +349,16 @@ machine!")
 
         """
 
-        pygame.image.save(self._surface, filename)
+        if self._open_gl:
+            size = self._surface.get_size()
+            ogl.glReadBuffer(ogl.GL_FRONT)
+            buffer = ogl.glReadPixels(0, 0, size[0], size[1], ogl.GL_RGB,
+                                      ogl.GL_UNSIGNED_BYTE)
+            ogl.glReadBuffer(ogl.GL_BACK)
+            surf = self._surface.copy()
+            content = pygame.image.fromstring(buffer, size, "RGB", True)
+            rect = pygame.Rect((0, 0), size)
+            surf.blit(content, rect)
+        else:
+            surf = self._surface
+        pygame.image.save(surf, filename)