+++ expyriment/misc/_miscellaneous.py
@@ -302,8 +302,32 @@ def find_font(font):
         return ""
 
 
+def get_display_info():
+    """Return information about the available display(s).
+
+    Return
+    ------
+    info : dict of dicts
+        the information for each available display in the form
+             "maximal_resolution": ...,  # highest possible or "native", tuple
+             "desktop_resolution": ...}  # might differ (e.g. scaled), tuple
+
+    """
+
+    pygame.display.init()
+    info = {}
+    for x in range(pygame.display.get_num_displays()):
+        info[x] = {
+             "maximal_resolution": pygame.display.list_modes(display=x)[0],
+             "desktop_resolution": pygame.display.get_desktop_sizes()[x]}
+
+    return info
+
+
 def get_monitor_resolution():
     """Returns the monitor resolution
+
+    DEPRECATED! Use get_display_info instead.
 
     Returns
     -------
@@ -312,13 +336,11 @@ def get_monitor_resolution():
 
     """
 
-    from .. import _internals
+    pygame.display.init()
     if _internals.active_exp.is_initialized:
-        return _internals.active_exp.screen.monitor_resolution
+        return _internals.active_exp.screen.display_resolution
     else:
-        pygame.display.init()
-        return (pygame.display.Info().current_w,
-                pygame.display.Info().current_h)
+        return get_display_info[0]["maximal_resolution"]
 
 
 def is_ipython_running():